+++ src/pages/expect/BiomarkerTest.js
@@ -1,13 +1,10 @@
-import React, { useState } from 'react'
-import { Button, Box, Container, Grid, Link, Typography, Stepper, Step, StepLabel, StepContent, Divider, Paper } from '@material-ui/core'
+import React from 'react'
+import { Box, Grid, Link, Typography, Stepper, Step, StepLabel, StepContent, Divider, Paper } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
-import { 
-  RemoveRounded as MinusIcon
-} from '@material-ui/icons'
 import { useTranslation } from 'react-i18next'
 import RenderContent from '../../components/utils/RenderContent'
 
-import FAQ from '../../components/FAQ/FAQ'
+import FAQs from '../../components/FAQ/FAQ_Wrapper'
 import TabPanel from '../../components/Tabs/TabPanel'
 
 const useStyles = makeStyles( theme => ({
@@ -84,16 +81,6 @@ const useStyles = makeStyles( theme => ({
     '& > a': {
       margin: theme.spacing(1,0)
     }
-  },
-  faq_title: {
-    display: 'flex',
-    '& h3': {
-      flexGrow: 1
-    }
-  },
-  faq_icon: {
-    marginLeft: -8,
-    marginRight: 4
   }
 }))
 
@@ -102,11 +89,6 @@ const BiomarkerTest = (props) => {
   const classes = useStyles()
   const { t, i18n } = useTranslation('testing')
   const faqs = i18n.getResourceBundle(i18n.languages[0],'testing').faqs
-  const [isExpanded, setIsExpanded] = useState(0)
-
-  const toggleAll = () => {
-    setIsExpanded(prev => prev + 1)
-  }
 
   return (
     <TabPanel
@@ -201,21 +183,8 @@ const BiomarkerTest = (props) => {
       </Box>
 
       {/* Frequently Asked Questions */}
-      <Container className="innerContainer">
-        <div className={classes.faq_title}>
-          <Typography variant="h3" component="h3">{t('faqs_title')}</Typography>
-          <Button variant="outlined" color="primary" onClick={toggleAll}><MinusIcon className={classes.faq_icon} /> {t('common:buttons.collapse')}</Button>
-        </div>
-        <Box mt={3}>
-          {faqs && faqs.map((faq, i) => <FAQ
-            key={i} 
-            title={faq.question}
-            desc={faq.answer}
-            expanded={isExpanded}
-          />)}
-        </Box>
-      </Container>
-      
+      <FAQs title={t('faqs_title')} faqs={faqs} />
+
     </TabPanel>
   )
 }
